third_party/WebKit/LayoutTests/http/tests/security/cookies/websocket/third-party-cookie-blocked-on-same-origin-websocket.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
if (window.testRunner) {
  // Block third-party cookies.
  testRunner.setBlockThirdPartyCookies(true);
}

async_test(t => {
  window.addEventListener(
      'message',
      t.step_func_done(event => {
        assert_equals(event.data, 'DONE: undefined');
      }),
      false);
}, "Test that third-party cookies are blocked for WebSockets even when the " +
    "WebSocket's URL's domain is an exact match for the domain of its " +
    "initiator.");
</script>
<!--
  The iframe is a remote origin. It uses a remote origin for the WebSocket.
-->
<iframe src="http://localhost:8000/security/cookies/websocket/resources/set-cookie.html?remote"></iframe>
